+++ src/snowflake/cli/api/commands/snow_typer.py
@@ -24,6 +24,7 @@ class SnowTyper(typer.Typer):
             **kwargs,
             context_settings=DEFAULT_CONTEXT_SETTINGS,
             pretty_exceptions_show_locals=False,
+            no_args_is_help=True,
         )
 
     @wraps(typer.Typer.command)