+++ src/snowflake/cli/app/cli_app.py
@@ -126,8 +126,9 @@ def _info_callback(value: bool):
         _exit_with_cleanup()
 
 
-@app.callback()
+@app.callback(invoke_without_command=True)
 def default(
+    ctx: typer.Context,
     version: bool = typer.Option(
         None,
         "--version",
@@ -211,6 +212,9 @@ def default(
     """
     Snowflake CLI tool for developers.
     """
+    if not ctx.invoked_subcommand:
+        typer.echo(ctx.get_help())
+
     setup_pycharm_remote_debugger_if_provided(
         pycharm_debug_library_path=pycharm_debug_library_path,
         pycharm_debug_server_host=pycharm_debug_server_host,