+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 ## Fixes and improvements
 * Adding `--image-name` option for image name argument in `spcs image-repository list-tags` for consistency with other commands.
 * Fixed errors during `spcs image-registry login` not being formatted correctly.
+* When invokes without commands help is displayed by default with list of available commands.
 
 # v2.1.0